GUI added
GUI work started. Option to choose graph, based on those in the user's
data folder. Also Node Count button in Other tab will load in that graph
to the buffers and perform a simple node count. This is done in another
thread to avoid locking up the GUI. This is shown to the user both with
a status message and loading bar.