GUI Updates & Refactoring
GUI now has the following updates:
* Can load existing buffers and perform a node count on them (as a
simple test)
* Can now convert a adjacency list into a series of buffer. Loads the
file from wherever, but saves the buffers in data/%name%/%name%.buf etc
* Have created helper functions in Launcher that call existing methods
in FastGraph. This is so that the GUI and the command line interface can
both directly call the same method
* These new methods are also the last places Exceptions are thrown.
Exceptions will travel up the stack to these methods before being thrown
one last time. The handling will depend on the usage - command line will
be an error and quit, while the GUI will is more graceful and displays
an error box. (Note the command line functions are yet to be
implemented)