Added in relabelling code
This now works. Identifies a simplied view of a family based on the
subgraphs in the famlies folder. As these are cliques, it picks two to
be parents, and labels the nodes with the same surname. Edges are set
with particular types based on the connections (parent, sibling,
married). Other edges and nodes are allocated randomly.

Also added another few conversion methods to Util.