Data deletion and node labelling complete
It is quicker to add every entry and then delete those which are not
needed than check each entry as it is added - it now takes 6s to do this
step compared to several hours.

Fixed bug in creating node and edge lists to remove where not all edges
were selected.

Nodes and Edges are now labelled.

Now takes 424s to delete as-skitter to the right size and 17s to relabel
it.