Motif finding is possible from the command line
Also added a few changes to the GUI motif finder.
Added a dummy class that overrides the methods of MotifTask, but has no
implementation. This avoids having if (mt != null) everywhere in the
ExactMotifFinder. Potentially in the future these methods could be
implemented, but by saving to a log file, or outputting to the screen.