Motif finding now works
Changes so that the FastGraphs for found motifs are no longer stored in
RAM, but on disk. This slows the process down, but allows it to run.

Also, a big list of all motifs is no longerr found then checked.
Instead, the motifs for each node are found and checked in turn. This
means a big set of FastGraphs is no longer needed in memory.

ExactIsomorphism objects are set to null once complete - this encourages
the garbage collector to retrieve them.